smb2: break out ioctl handling
