+++ rust/src/smb/mod.rs
@@ -27,6 +27,7 @@ pub mod smb1;
 pub mod smb1_session;
 pub mod smb2;
 pub mod smb2_session;
+pub mod smb2_ioctl;
 pub mod smb3;
 pub mod dcerpc;
 pub mod session;
